Moved subgraph and metadata APIs to src
