+++ src/api/index.ts
@@ -1,3 +1,3 @@
 export * from './cow'
-export * from './cow-subgraph'
-export * from './metadata'
+export * from '../cow-subgraph'
+export * from '../metadata'
